refactor: Rename Redis Pub/Sub
